pyproject.toml
[project]
name = "downloader-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2025.11.12",
]
